# swagger_doc.yaml
swagger: "2.0"
info:
  title: "My API"
  description: "API Documentation"
  version: "1.0.0"
host: "localhost:5000"
basePath: "/"
schemes:
  - "http"
paths:
  /login:
    post:
      summary: "User Login Endpoint"
      description: "Authenticate users and generate access token"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      parameters:
        - name: "username"
          in: "formData"
          type: "string"
          required: true
          description: "The username for login"
        - name: "password"
          in: "formData"
          type: "string"
          required: true
          description: "The password for login"
      responses:
        200:
          description: "User logged in successfully"
          examples:
            access_token: "JWT access token"
        401:
          description: "Invalid username or password"
  /data:
    get:
      summary: "Get All Data Endpoint"
      description: "Retrieve all data"
      produces:
        - "application/json"
      security:
        - JWT: []
      responses:
        200:
          description: "A list of all data retrieved successfully"
          schema:
            type: "array"
            items:
              type: "object"
              properties:
                id:
                  type: "string"
                name:
                  type: "string"
                value:
                  type: "integer"
    post:
      summary: "Add Data Endpoint"
      description: "Add new data"
      consumes:
        - "application/json"
      produces:
        - "application/json"
      security:
        - JWT: []
      parameters:
        - name: "body"
          in: "body"
          required: true
          schema:
            type: "object"
            properties:
              name:
                type: "string"
              value:
                type: "integer"
      responses:
        201:
          description: "Data added successfully"
    /{id}:
      put:
        summary: "Update Data Endpoint"
        description: "Update existing data"
        consumes:
          - "application/json"
        produces:
          - "application/json"
        security:
          - JWT: []
        parameters:
          - name: "id"
            in: "path"
            required: true
            type: "string"
            description: "ID of the data to be updated"
          - name: "body"
            in: "body"
            required: true
            schema:
              type: "object"
              properties:
                name:
                  type: "string"
                value:
                  type: "integer"
        responses:
          200:
            description: "Data updated successfully"
      delete:
        summary: "Delete Data Endpoint"
        description: "Delete existing data"
        produces:
          - "application/json"
        security:
          - JWT: []
        parameters:
          - name: "id"
            in: "path"
            required: true
            type: "string"
            description: "ID of the data to be deleted"
        responses:
          200:
            description: "Data deleted successfully"
securityDefinitions:
  JWT:
    type: "apiKey"
    name: "Authorization"
    in: "header"